# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      name:
        maxLength: 200
        minLength: 3
        type: string
      price:
        type: number
      sku:
        maxLength: 50
        minLength: 3
        type: string
      stock_quantity:
        minimum: 0
        type: integer
    required:
    - category_id
    - sku
    - stock_quantity
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      expires_in:
        type: integer
      message:
        type: string
      remaining_tries:
        type: integer
      retry_after:
        type: integer
      success:
        type: boolean
      token:
        type: string
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      status:
        type: string
      stock_quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  models.UpdateProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      name:
        maxLength: 200
        minLength: 3
        type: string
      price:
        type: number
      status:
        enum:
        - active
        - inactive
        - discontinued
        type: string
      stock_quantity:
        minimum: 0
        type: integer
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      update_at:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - username
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
host: localhost:8085
info:
  contact:
    email: aaravmahajan2003@gmail.com
    name: Aarav Mahajan
    url: https://github.com/aaravmahajanofficial
  description: This is the API server for the Scalable E-commerce Platform. It provides
    endpoints for managing users, products, carts, orders, payments, and notifications.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scalable E-commerce Platform API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Retrieves a paginated list of available products. Requires authentication.
      parameters:
      - description: 'Page number for pagination (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Number of items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of products
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponse'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/models.Product'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List products with pagination
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Adds a new product to the catalog. Requires authentication.
      parameters:
      - description: Product Creation Details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created product
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Validation error or invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    get:
      description: Retrieves details for a specific product using its ID. Requires
        authentication.
      parameters:
      - description: Product ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved product
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid product ID format
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Updates details for an existing product using its ID. Requires
        authentication.
      parameters:
      - description: Product ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Product Update Details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated product
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid product ID format or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product by ID
      tags:
      - Products
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token upon successful login.
      parameters:
      - description: User Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login, includes JWT token
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Validation error or invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too many login attempts
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Log in a user
      tags:
      - Users
  /users/profile:
    get:
      description: Retrieves the profile information for the currently authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user profile
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Authentication required (invalid or missing token)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided details.
      parameters:
      - description: User Registration Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Validation error or invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: User with email already exists
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"